+++ Navigation/ProfileViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ProfileViewController.swift
+//  Navigation
+//
+//  Created by Мамуля on 13.06.2023.
+//
+
+import UIKit
+
+class ProfileViewController: UIViewController {
+        
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+        view.backgroundColor = .systemYellow
+    }
+        
+        
+}
